+++ src/main/java/aoc/y2019/Day05Solution.java
@@ -87,7 +87,7 @@ public class Day05Solution {
     @Test
     public void part1() {
         Memory memory = Memory.init(input(this));
-        final var computer = new Computer(memory, new IO(ImmutableList.of(1)));
+        final var computer = new Computer(memory, new IO(ImmutableList.of(1L)));
         computer.runToCompletion();
 
         final var output = computer.io().output();
@@ -116,7 +116,7 @@ public class Day05Solution {
         assertOutput(4283952, input(this), 5);
     }
 
-    private void assertOutput(int expectedOutput, Input program, int input) {
+    private void assertOutput(long expectedOutput, Input program, long input) {
         var computer = new Computer(Memory.init(program), new IO(ImmutableList.of(input)));
         computer.runToCompletion();
         assertEquals(List.of(expectedOutput), computer.io().output());
